LocationMapper added
