+++ add-location.http
@@ -5,6 +5,6 @@ Content-Type: application/json
   "city": "NY",
   "region": "pomorskie",
   "country": "Polska",
-  "latitude": 0.0,
+  "latitude": 1.0,
   "longitude": 0.0
 }
