+++ src/main/java/com/mkgn/weatherservice/location/NewLocationResponse.java
@@ -1,12 +1,14 @@
 package com.mkgn.weatherservice.location;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
+@Builder
 class NewLocationResponse {
     private String id;
     private String city;
